partial_dtb_gen: Add initial version
Added the initial version of the tool, which takes as input a
complete Linux DTB file, a Template DTB/DTS file and a node which
to transfer from the former to the latter, updating its properties
with respect to the xen passthrough format.

Also, the script will add all the referenced memory-regions of the
node to the output DTS file. The output of the script is a
"passthrough.dts" file, which will contain the characteristics
mentioned before.

Additionally, I added an example for the Template DT file. It
needs to be compiled first to a .dtb format to be loaded into
the script. The yocto recipes where the tool will be used will
take care of that.

Issue: ALB-7428
Signed-off-by: Andrei Cherechesu <andrei.cherechesu@nxp.com>